[TASK] Merge `Csv/AbstractListView` into `AbstractRegistrationListView`
Fixes #3697
